+++ src/Common/buildMQTTDeviceData.ts
@@ -4,13 +4,15 @@ import { safeId } from '@utils/safeId';
 type Device = { friendlyName: string; name: string; address: number | string };
 
 export const buildMQTTDeviceData = ({ friendlyName, name, address }: Device, manufacturer: string): IDeviceData => {
+  const deviceTopic = `${safeId(manufacturer)}/${safeId(address.toString())}`;
   return {
-    deviceTopic: `${safeId(manufacturer)}/${safeId(address.toString())}`,
+    deviceTopic,
     device: {
       ids: [`${address}`],
       name: friendlyName,
       mf: manufacturer,
       mdl: name,
     },
+    availabilityTopic: `${deviceTopic}/bleConnection/status`,
   };
 };