+++ src/Richmat/Nordic/controllerBuilder.ts
@@ -2,7 +2,11 @@ import { IDeviceData } from '@ha/IDeviceData';
 import { BLEController } from 'BLE/BLEController';
 import { IBLEDevice } from 'ESPHome/types/IBLEDevice';
 
-export const controllerBuilder = async (deviceData: IDeviceData, bleDevice: IBLEDevice) => {
+export const controllerBuilder = async (
+  deviceData: IDeviceData,
+  bleDevice: IBLEDevice,
+  stayConnected: boolean | undefined
+) => {
   const { getCharacteristic } = bleDevice;
 
   const characteristic = await getCharacteristic(
@@ -11,5 +15,5 @@ export const controllerBuilder = async (deviceData: IDeviceData, bleDevice: IBLE
   );
   if (!characteristic) return undefined;
 
-  return new BLEController(deviceData, bleDevice, characteristic.handle, (byte: number) => [byte]);
+  return new BLEController(deviceData, bleDevice, characteristic.handle, (byte: number) => [byte], {}, stayConnected);
 };