+++ src/Richmat/WiLinke/controllerBuilder.ts
@@ -6,13 +6,17 @@ import { supportedBeds } from './supportedBeds';
 
 const buildCommand = (command: number) => [110, 1, 0, command, command + 111];
 
-export const controllerBuilder = async (deviceData: IDeviceData, bleDevice: IBLEDevice) => {
+export const controllerBuilder = async (
+  deviceData: IDeviceData,
+  bleDevice: IBLEDevice,
+  stayConnected: boolean | undefined
+) => {
   const { name, getCharacteristic } = bleDevice;
 
   for (const { serviceUuid, writeCharacteristicUuid } of supportedBeds) {
     const characteristic = await getCharacteristic(serviceUuid, writeCharacteristicUuid, false);
     if (!characteristic) continue;
-    return new BLEController(deviceData, bleDevice, characteristic.handle, buildCommand);
+    return new BLEController(deviceData, bleDevice, characteristic.handle, buildCommand, {}, stayConnected);
   }
 
   logInfo('[Richmat] Could not find any supported services/characteristics for device:', name);