+++ src/Solace/solace.ts
@@ -2,6 +2,7 @@ import { IMQTTConnection } from '@mqtt/IMQTTConnection';
 import { buildDictionary } from '@utils/buildDictionary';
 import { logError, logInfo } from '@utils/logger';
 import { BLEController } from 'BLE/BLEController';
+import { setupConnectionAvailability } from 'BLE/setupConnectionAvailability';
 import { setupDeviceInfoSensor } from 'BLE/setupDeviceInfoSensor';
 import { buildMQTTDeviceData } from 'Common/buildMQTTDeviceData';
 import { IESPConnection } from 'ESPHome/IESPConnection';
@@ -46,6 +47,8 @@ export const solace = async (mqtt: IMQTTConnection, esphome: IESPConnection) => 
       {},
       true
     );
+    if (controller.isPersistentConnection) setupConnectionAvailability(mqtt, bleDevice, deviceData);
+
     logInfo('[Solace] Setting up entities for device:', name);
     setupPresetButtons(mqtt, controller);
     setupMotorEntities(mqtt, controller);